@@ -1,6 +1,6 @@
 package main
 
-import "github.com/elahe-dastan/goURL/internal/cmd"
+import "github.com/1995parham-learning/gourl/internal/cmd"
 
 func main() {
 	cmd.Execute()
